backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
joblib
requests
flask-sqlalchemy
pyjwt
